NIO: swap selector implementations (NFC)
In most of the code, Linux is treated as the fallback case.  This
adjusts the selector implementation to follow the same pattern, checking
if on Darwin and using kqueue (which would also be available on F/BSD)
and otherwise falling back to `epoll` (assuming Linux).  This
homogeneous behaviour makes it easier to reason about when trying to add
support for additional platforms.