Shorthand way of setting options in server bootstrap.
Motivation:

Less typing is good.

Modifications:

Add a new collection of types to provide shorthand channel options.
A new method on ServerBootstrap to set an array of these shorthand options.
A test case that long hand does the same as short hand.
Change to HTTPServer to demonstrate the change.

Result:

There will be an new way of setting options involving less typing.