Add preconditionNotInEventLoop() and assertNotInEventLoop() methods to EventLoop.
Motivation:

As it is useful to have a precondition to assert that calling code *is* on the given EventLoop, it is also useful to assert that calling code is *not* on it.

Modifications:

Add two new methods to EventLoop, one of them as a protocol method, both with default implementations.

Result:

It is now possible to canonically assert that a given execution path is *not* executing on a given event loop (including providing an explanatory message, as with the earlier changes to the counterpart methods). This provides better parity with dispatchPrecondition() and enables replacing direct access to `EventLoop.inEventLoop` with a direct statement of intent providing guaranteed semantics.