Make use of the new precondition/assertNotInEventLoop() methods in appropriate locations.
Motivation:

This is the purpose for which the new methods were intended: To replace handwritten conditions with direct statements of assertion intent.

Modifications:

Replaced various uses of precondition(!eventLoop.inEventLoop) and similar assert()s with the respective *notInEventLoop methods. In the case of the usage in EventLoopFuture.wait(), ensure the explainer message is provided on precondition failure.

Result:

The modified code is clearer, explicitly declares intent, can benefit from customizations provided by individual event loop types (such as the use of dispatchPrecondition() in NIOTSEventLoop), and can be more easily refactored.