Make SocketAddress Hashable.
Motivation:

In general there's no good justification for `SocketAddress` not to be
`Hashable`: it's a well-formed value type that can easily have a
clearly-defined version of hashability defined. In particular, we have
a clear definition of equatability for this type already present in the
code.

This makes SocketAddress more broadly useful in a range of
circumstances.

Modifications:

- Defined `Hashable` for `SocketAddress`.
- Fixed broken `Equatable` implementation for `SocketAddress`.

Result:

Users can obtain the hash value of `SocketAddress`.