remove XCTest expectations with timeouts (#1539)
Motivation:

The XCTest expectations caused problems when they time out. When they
time out, the execution continues which in the tests in question then
shut down the EventLoop. Shutting down the event loop meant that other
things don't happen which caused the test to hang. The XCTest
expectations however tell us about the failure at the end of the test
(which we never reach).

Modifications:

Remove the XCTest expectations with timeouts.

Result:

Even slow CI should pass this test correctly.

Co-authored-by: Cory Benfield <lukasa@apple.com>