NIO: initial pass over iov handling for Windows
This partially cleans up the iovec based message handling.  Create a new
typealias for `IOVector` which changes the type for the underlying
implementation's scatter/gather IO type.  This only covers the cases
where there is an implemented equivalent.  There is still work to be
done to cover some of the other cases.